app/requirements.txt
celery==4.1.0
Django==1.11.10
djangorestframework==3.7.7
psycopg2==2.7.3.2